requirements.txt
# Configuration & CLI
hydra-core==1.3.2
omegaconf==2.3.0
typer==0.9.0
pydantic==2.5.2

# Experiment tracking & monitoring
mlflow==2.9.2
tensorboard==2.15.1
wandb==0.16.1

# Development tools
jupyter==1.1.1
ipywidgets==8.1.5
pytest==7.4.3
pytest-cov==4.1.0

# Code quality
black==23.12.1
isort==5.13.2
flake8==6.1.0
mypy==1.8.0
pre-commit==3.6.0

# Visualization
plotly==5.18.0
seaborn==0.13.0
matplotlib==3.8.2

# Logging & utilities
rich==13.7.0
loguru==0.7.2
tqdm==4.66.1

# Hyperparameter optimization
optuna==3.5.0

# Model interpretation
shap==0.44.0

# Profiling
py-spy==0.3.14
memory-profiler==0.61.0

# Parallel processing
joblib==1.3.2
